requirements.txt
flask==3.0.2
werkzeug~=3.0.1
python-dotenv==1.0.1
click~=8.1.7
Flask-WTF~=1.2.1
WTForms~=3.1.1
wtforms-sqlalchemy~=0.4.1
email-validator~=2.2.0
alchemical~=1.0.1
SQLAlchemy==2.0.27
pymysql~=1.1.0
pytz~=2024.1
flask-mail~=0.10.0